+++ src/app/exports/modal/modal-window.component.ts
@@ -11,6 +11,7 @@ import { ModalContentComponent } from './modal-content.component';
 import { ModalOptions } from './modal-options.model';
 import { ModalDismissReasons } from './modal-dismiss-reasons.model';
 import { trigger, state, style, transition, animate } from '@angular/animations';
+import { timer } from 'rxjs/observable/timer';
 
 @Component({
   selector: 're-modal-window',
@@ -43,13 +44,18 @@ export class ModalWindowComponent {
 
   open() {
     this.isOpen = true;
-    this.animateState = 'in';
+    if (this.modalOptions.animation) {
+      this.animateState = 'in';
+    }
   }
 
   close() {
     this.isOpen = false;
-    this.animateState = 'void';
-    return this.animationDone;
+    if (this.modalOptions.animation) {
+      this.animateState = 'void';
+      return this.animationDone;
+    }
+    return timer(0);
   }